requirements.txt
django>=1.11
psycopg2==2.7.4
django-extensions
Pillow
ipdb
djangorestframework
uwsgi
